+++ examples/script_auth_friend_list.py
@@ -14,7 +14,7 @@ import sys
 
 def main():
     """Provide the program's entry point when directly executed."""
-    authenticator = prawcore.Authenticator(
+    authenticator = prawcore.TrustedAuthenticator(
         prawcore.Requestor('prawcore_script_auth_example'),
         os.environ['PRAWCORE_CLIENT_ID'],
         os.environ['PRAWCORE_CLIENT_SECRET'])