+++ prawcore/__init__.py
@@ -1,8 +1,9 @@
 """prawcore: Low-level communication layer for PRAW 4+."""
 
 import logging
-from .auth import (Authenticator, Authorizer, DeviceIDAuthorizer,  # noqa
-                   ReadOnlyAuthorizer, ScriptAuthorizer)
+from .auth import (Authorizer, DeviceIDAuthorizer, ReadOnlyAuthorizer,  # NOQA
+                   ScriptAuthorizer, TrustedAuthenticator,
+                   UntrustedAuthenticator)
 from .const import __version__  # noqa
 from .exceptions import *  # noqa
 from .requestor import Requestor  # noqa